deploy: to firebase

+++ client/src/stores/counter.js
@@ -1,10 +1,12 @@
 import { defineStore } from 'pinia'
 import axios from 'axios'
+import Swal from 'sweetalert2'
 
 export const useFunctionStore = defineStore('function', {
   state: () => ({
     baseUrl: `https://nusantaralounge-production.up.railway.app`,
     currentUser: "",
+    role: localStorage.role,
 
     //dashboard data
     heroData: [],
@@ -17,6 +19,9 @@ export const useFunctionStore = defineStore('function', {
     podcastAllData: [],
     podcastDetailData: [],
 
+    //favorite data
+    playlistAllData: [],
+
     isLogin: localStorage.access_token
   }),
 
@@ -27,6 +32,52 @@ export const useFunctionStore = defineStore('function', {
 
 
   actions: {
+
+    async handleGoogleSign(response) {
+      try {
+        const googlesign = await axios({
+          url: this.baseUrl + '/user/google-sign-in',
+          // url: `http://localhost:3000` + '/user/google-sign-in',
+          method: "POST",
+          headers: {
+            "google-oauth-token": response.credential,
+          },
+        });
+        localStorage.access_token = googlesign.data.access_token;
+        localStorage.currentUser = googlesign.data.email;
+        localStorage.role = googlesign.data.role
+
+        let timerInterval = 5000
+        Swal.fire({
+          title: 'Successfly Login',
+          html: 'going to dashboard in <b></b> milliseconds.',
+          timer: 2000,
+          timerProgressBar: true,
+          didOpen: () => {
+            Swal.showLoading()
+            const b = Swal.getHtmlContainer().querySelector('b')
+            timerInterval = setInterval(() => {
+              b.textContent = Swal.getTimerLeft()
+            }, 100)
+          },
+          willClose: () => {
+            clearInterval(timerInterval)
+          }
+        }).then((result) => {
+          /* Read more about handling dismissals below */
+          if (result.dismiss === Swal.DismissReason.timer) {
+            console.log('I was closed by the timer')
+            this.router.push("/")
+          }
+        })
+
+      } catch (error) {
+        Swal.fire(error.response.data.message)
+
+        // console.log(error);
+      }
+    },
+
 
     async handleLogin(dataLogin) {
       try {
@@ -40,21 +91,59 @@ export const useFunctionStore = defineStore('function', {
         })
         localStorage.access_token = data.access_token
         localStorage.currentUser = data.email
-        this.router.push('/')
+        localStorage.role = data.role
+
+        let timerInterval = 5000
+        Swal.fire({
+          title: 'Successful Login',
+          html: 'going to dashboard in <b></b> milliseconds.',
+          timer: 2000,
+          timerProgressBar: true,
+          didOpen: () => {
+            Swal.showLoading()
+            const b = Swal.getHtmlContainer().querySelector('b')
+            timerInterval = setInterval(() => {
+              b.textContent = Swal.getTimerLeft()
+            }, 100)
+          },
+          willClose: () => {
+            clearInterval(timerInterval)
+          }
+        }).then((result) => {
+          /* Read more about handling dismissals below */
+          if (result.dismiss === Swal.DismissReason.timer) {
+            console.log('I was closed by the timer')
+            this.router.push("/")
+          }
+        })
       } catch (error) {
-        console.log(error.response.data.message);
+        Swal.fire(error.response.data.message)
+        // console.log(error.response.data.message);
       }
     },
 
     handleLogout() {
-      localStorage.clear()
-      this.currentUser = ""
-      this.isLogin = false
-      this.router.push('/login')
+      Swal.fire({
+        title: 'Are you sure want to logout?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Logout'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire(
+            'Logging out....',
+            localStorage.clear(),
+            this.currentUser = "",
+            this.isLogin = false,
+            this.router.push('/login')
+          )
+        }
+      })
     },
 
     async handleRegister(dataLogin) {
-      console.log(dataLogin);
       try {
         const { data } = await axios({
           url: this.baseUrl + `/user/register`,
@@ -64,9 +153,24 @@ export const useFunctionStore = defineStore('function', {
             password: dataLogin.password
           }
         })
-        console.log(data);
+        Swal.fire({
+          title: 'Submit Registration?',
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Logout'
+        }).then((result) => {
+          if (result.isConfirmed) {
+            Swal.fire(
+              'Success registered',
+            )
+          }
+        })
       } catch (error) {
-        console.log(error.response.data.message);
+        Swal.fire(error.response.data.message)
+
+        // console.log(error.response.data.message);
       }
     },
 
@@ -80,15 +184,50 @@ export const useFunctionStore = defineStore('function', {
         localStorage.url = url
         console.log(this.podcastDetailData);
       } catch (error) {
-        console.log(error);
+        Swal.fire(error.response.data.message)
+        // console.log(error);
       }
     },
+
+    async handlePayment() {
+      try {
+        const { data } = await axios({
+          url: this.baseUrl + '/user/generate-midtrans-token',
+          method: `POST`,
+          headers: { access_token: localStorage.access_token }
+        })
+        let access = window.snap.pay(data.token, {
+          onSuccess: function (result) {
+            /* You may add your own implementation here */
+            alert("payment success!"); console.log(result);
+            +
+              axios({
+                url: 'https://nusantaralounge-production.up.railway.app/user/subscription',
+                method: `PATCH`,
+                headers: { access_token: localStorage.access_token }
+              })
+            localStorage.role = `Premium`
+            Swal.fire('Congrats you are a premium member Now')
+          }
+        })
+        
+        // this.router.push('/')
+
+      } catch (error) {
+        Swal.fire(error.response.data.message)
+
+        // console.log(error);
+      }
+    },
+
+
 
     async fetchAll() {
       this.currentUser = localStorage.currentUserx
       this.fetchHeroData()
       this.fetchPodcast()
       this.fetchGames()
+      this.fetchFavorite()
     },
 
     async fetchHeroData() {
@@ -99,7 +238,9 @@ export const useFunctionStore = defineStore('function', {
         this.heroData = { big: data.articles[0], small: data.articles.slice(1, 5) }
         this.sideData = data.articles.slice(6, 16)
       } catch (error) {
-        console.log(error);
+        Swal.fire(error.response.data.message)
+
+        // console.log(error);
       }
     },
 
@@ -112,7 +253,9 @@ export const useFunctionStore = defineStore('function', {
         this.podcastAllData = data.podcast
         console.log(this.podcastData);
       } catch (error) {
-        console.log(error);
+        Swal.fire(error.response.data.message)
+
+        // console.log(error);
       }
     },
 
@@ -123,7 +266,86 @@ export const useFunctionStore = defineStore('function', {
         })
         this.gamesData = data.slice(0, 9)
       } catch (error) {
-        console.log(error);
+        Swal.fire(error.response.data.message)
+
+        // console.log(error);
+      }
+    },
+
+    async fetchFavorite() {
+      try {
+        const { data } = await axios({
+          url: this.baseUrl + `/playlist`,
+          headers: { access_token: localStorage.access_token }
+        })
+        this.playlistAllData = data
+        console.log(this.playlistAllData);
+      } catch (error) {
+        Swal.fire(error.response.data.message)
+
+        // console.log(error);
+      }
+    },
+
+    async addToPlaylist(dataPodcast) {
+      console.log(dataPodcast.title, '<<<<<<<<<<<<<<');
+      try {
+
+        const { data } = await axios({
+          url: this.baseUrl + `/playlist`,
+          method: `POST`,
+          data: {
+            title: dataPodcast.title,
+            image: dataPodcast.image,
+            audio: dataPodcast.audio,
+            post_content: dataPodcast.post_content,
+            published_at: dataPodcast.published_at,
+          },
+          headers: { access_token: localStorage.access_token }
+        })
+        Swal.fire('Successfuly added to playlist')
+        this.playlistAllData = data
+      } catch (error) {
+        Swal.fire(error.response.data.message)
+
+        // console.log(error);
+      }
+    },
+
+    async deletePlaylist(dataPodcast) {
+      console.log(dataPodcast, '<<<<<<<<<<<<<<');
+      try {
+        Swal.fire({
+          title: 'Are you sure?',
+          text: "You won't be able to revert this!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes, remove from playlist!'
+        }).then((result) => {
+          if (result.isConfirmed) {
+            Swal.fire(
+              'Remove from favorite',
+            )
+            const { data } = axios({
+              url: this.baseUrl + `/playlist`,
+              method: `delete`,
+              headers: { access_token: localStorage.access_token },
+              data: {
+                PlaylistId: dataPodcast.PlaylistId,
+                UserPlaylistId: dataPodcast.UserPlaylistId,
+              }
+            })
+            this.fetchFavorite()
+            console.log(data);
+
+          }
+        })
+      } catch (error) {
+        Swal.fire(error.response.data.message)
+
+        // console.log(error);
       }
     },
